+++ docs/issue-031/PLAN.md
@@ -158,29 +158,34 @@ test('改行コードを無視した文字列置換機能のe2eテスト', async
 });
 ```
 
-## 実装手順
+## 実装手順と現在の状況
 
 ### ステップ 1: Value Object作成
-- [ ] `NormalizedString.ts` 作成
-- [ ] `NormalizedString.test.ts` 作成・テスト実装
+- [x] `NormalizedString.ts` 作成 ✓ 実装完了
+- [x] `NormalizedString.test.ts` 作成・テスト実装 ✓ 実装完了
 
 ### ステップ 2: HtmlContent修正
-- [ ] `HtmlContent.ts` の `replaceWith` メソッド修正
-- [ ] 既存のユニットテスト修正・新規テスト追加
+- [x] `HtmlContent.ts` の `replaceWith` メソッド修正 ✓ 実装完了
+  - 正規化文字列による複数回の置換ロジック実装済み
+  - `findNormalizedMatch` および `findActualRangeInString` メソッド実装済み
+- [x] 既存のユニットテスト修正・新規テスト追加 ✓ 実装完了
 
 ### ステップ 3: e2eテスト追加
-- [ ] `ignore-crlf-replace.spec.ts` 作成・実装
+- [x] `ignore-crlf-replace.spec.ts` 作成・実装 ✓ 実装完了
 
 ### ステップ 4: 統合テスト
-- [ ] 全てのユニットテストが通ることを確認
-- [ ] 全てのe2eテストが通ることを確認
-- [ ] 既存機能に影響がないことを確認
+- [x] 全てのユニットテストが通ることを確認 ✓ 完了（前回セッションで27テストが成功）
+- [x] 全てのe2eテストが通ることを確認 ✓ 完了
+- [x] 既存機能に影響がないことを確認 ✓ 完了
+
+### ステップ 5: コードクリーンアップ（残作業）
+- [ ] デバッグログの削除（`HtmlContent.ts`の`console.log`文を削除）
 
 ## 受け入れ条件チェック
-- [ ] 上記ユースケースのe2eテストが追加され、パスするように実装が変更されている
-- [ ] そのe2eテストが通ること
-- [ ] 既存のe2eテストが通ること
-- [ ] 変更箇所については、`CODING_STYLE.md` の `# オブジェクト指向ルール` に従っていること
+- [x] 上記ユースケースのe2eテストが追加され、パスするように実装が変更されている ✓
+- [x] そのe2eテストが通ること ✓
+- [x] 既存のe2eテストが通ること ✓
+- [x] 変更箇所については、`CODING_STYLE.md` の `# オブジェクト指向ルール` に従っていること ✓
 
 ## 技術的制約・考慮事項
 - オブジェクト指向ルール（ThoughtWorksアンソロジー）への準拠
