--- host-frontend-root/frontend-src-root/src/domain/value-objects/WorkingHtml.ts
@@ -1,80 +1,0 @@
-import { TextRange } from './TextRange';
-import { RewriteRule } from '../entities/RewriteRule';
-
-/**
- * 置換処理中のHTML文字列を表すValue Object
- * HTML文字列の操作に必要なメソッドをカプセル化する
- */
-export class WorkingHtml {
-  private readonly value: string;
-  private readonly rule: RewriteRule;
-
-  constructor(value: string, rule: RewriteRule) {
-    this.value = value;
-    this.rule = rule;
-  }
-
-  /**
-   * 文字列を取得
-   */
-  toString(): string {
-    return this.value;
-  }
-
-  /**
-   * 指定した範囲を新しい文字列で置換した新しいWorkingHtmlを返す
-   * @param matchRange 置換対象の範囲
-   * @param newString 置換する新しい文字列
-   * @returns 置換後の新しいWorkingHtmlオブジェクト
-   */
-  replaceRange(matchRange: TextRange, newString: string): WorkingHtml {
-    const updatedHtml = this.value.substring(0, matchRange.start) + 
-                       newString + 
-                       this.value.substring(matchRange.end);
-    return new WorkingHtml(updatedHtml, this.rule);
-  }
-
-  /**
-   * RewriteRuleに基づいて最初に見つかったマッチ箇所を置換
-   * 冗長化された正規表現を使用して元のHTMLから直接マッチ位置を取得
-   * @returns 置換後の新しいWorkingHtmlオブジェクト
-   */
-  replaceByNormalizedPosition(): WorkingHtml {
-    // rule.oldStringを冗長化した正規表現を作成
-    const redundantPattern = this.createRedundantPattern(this.rule.oldString);
-    
-    // 正規化前のHTMLから直接マッチ位置を取得
-    const match = this.value.match(redundantPattern);
-    if (!match) {
-      // マッチしない場合は元のHTMLをそのまま返す
-      return this;
-    }
-    
-    const start = match.index!;
-    const end = start + match[0].length;
-    
-    // TextRangeを内部で作成して、replaceRangeメソッドを使用して置換
-    const range = new TextRange(start, end);
-    return this.replaceRange(range, this.rule.newString);
-  }
-
-  /**
-   * rule.oldStringを冗長化した正規表現パターンを作成
-   * `<` → `\\s*<`、`>` → `>\\s*` の変換を行う
-   * @param oldString 元の検索文字列
-   * @returns 冗長化された正規表現
-   */
-  private createRedundantPattern(oldString: string): RegExp {
-    // 正規表現の特殊文字をエスケープ
-    const escaped = oldString.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    
-    // `<` → `\\s*<`、`>` → `>\\s*` の変換を適用
-    const redundantPattern = escaped
-      .replace(/</g, '\\s*<')
-      .replace(/>/g, '>\\s*');
-    
-    return new RegExp(redundantPattern);
-  }
-
-
-}
